requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
faiss-cpu
openai
llama-cpp-python
coqui-tts
faster-whisper
loguru
twilio
httpx
python-multipart